requirements.txt
python-dotenv
schedule